split "linear" benchmark based on space usage
Benchmarks are divided into 4 categories:

1) using O(1) space
2) using O(n) Heap
3) using O(n) stack
4) using O(n) heap and stack

This allows us to run the benchmark executables with restricted heap
and stack memory limites so that we can catch space leak issues or high
space usage issues quickly.
