+++ benchmark/Common.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE CPP #-}
+#if __GLASGOW_HASKELL__ >= 800
+{-# OPTIONS_GHC -Wno-orphans #-}
+#endif
+
 -- |
 -- Module      : Main
 -- Copyright   : (c) 2019 Composewell Technologies
@@ -5,18 +10,106 @@
 -- License     : BSD3
 -- Maintainer  : streamly@composewell.com
 
-module Common (parseCLIOpts) where
+module Common
+    ( parseCLIOpts
 
+    , benchIOSink1
+    , benchPure
+    , benchPureSink1
+
+    , benchIOSrc1
+    , benchPureSrc
+
+    , mkString
+    , mkList
+    , mkListString
+
+    , defaultStreamSize
+    , limitStreamSize
+    )
+where
+
+import Control.DeepSeq (NFData(..))
 import Control.Exception (evaluate)
 import Control.Monad (when)
+import Data.Functor.Identity (Identity, runIdentity)
 import Data.List (scanl')
 import Data.Maybe (catMaybes)
 import System.Console.GetOpt
        (OptDescr(..), ArgDescr(..), ArgOrder(..), getOpt')
 import System.Environment (getArgs, lookupEnv, setEnv)
 import Text.Read (readMaybe)
+import System.Random (randomRIO)
 
+import qualified Streamly.Prelude as S
+
+import Streamly
 import Gauge
+
+-------------------------------------------------------------------------------
+-- Benchmarking utilities
+-------------------------------------------------------------------------------
+
+#if !MIN_VERSION_deepseq(1,4,3)
+instance NFData Ordering where rnf = (`seq` ())
+#endif
+
+-- XXX once we convert all the functions to use this we can rename this to
+-- benchIOSink
+{-# INLINE benchIOSink1 #-}
+benchIOSink1 :: NFData b => String -> (Int -> IO b) -> Benchmark
+benchIOSink1 name f = bench name $ nfIO $ randomRIO (1,1) >>= f
+
+{-# INLINE benchIOSrc1 #-}
+benchIOSrc1 :: String -> (Int -> IO ()) -> Benchmark
+benchIOSrc1 name f = bench name $ nfIO $ randomRIO (1,1) >>= f
+
+{-# INLINE benchPure #-}
+benchPure :: NFData b => String -> (Int -> a) -> (a -> b) -> Benchmark
+benchPure name src f = bench name $ nfIO $ randomRIO (1,1) >>= return . f . src
+
+-- XXX once we convert all the functions to use this we can rename this to
+-- benchPureSink
+{-# INLINE benchPureSink1 #-}
+benchPureSink1 :: NFData b => String -> (Int -> Identity b) -> Benchmark
+benchPureSink1 name f =
+    bench name $ nfIO $ randomRIO (1,1) >>= return . runIdentity . f
+
+{-# INLINE benchPureSrc #-}
+benchPureSrc :: String -> (Int -> SerialT Identity a) -> Benchmark
+benchPureSrc name src = benchPure name src (runIdentity . S.drain)
+
+-------------------------------------------------------------------------------
+-- String/List generation for read instances
+-------------------------------------------------------------------------------
+
+{-# INLINABLE mkString #-}
+mkString :: Int -> String
+mkString value = "fromList [1" ++ concat (replicate value ",1") ++ "]"
+
+{-# INLINABLE mkListString #-}
+mkListString :: Int -> String
+mkListString value = "[1" ++ concat (replicate value ",1") ++ "]"
+
+{-# INLINABLE mkList #-}
+mkList :: Int -> [Int]
+mkList value = [1..value]
+
+-------------------------------------------------------------------------------
+-- Stream size
+-------------------------------------------------------------------------------
+
+defaultStreamSize :: Int
+defaultStreamSize = 100000
+
+limitStreamSize :: Int -> IO Int
+limitStreamSize value = do
+    let val = min value defaultStreamSize
+    when (val /= value) $
+        putStrLn $ "Limiting stream size to "
+                   ++ show defaultStreamSize
+                   ++ " for non O(1) space operations"
+    return val
 
 -------------------------------------------------------------------------------
 -- Parse custom CLI options
@@ -55,7 +148,7 @@ deleteOptArgs (Just prev, _) opt =
     else (Just opt, Just opt)
 
 parseCLIOpts :: Int -> IO (Int, Config, [String])
-parseCLIOpts defaultStreamSize = do
+parseCLIOpts defStreamSize = do
     args <- getArgs
 
     -- Parse custom options
@@ -83,13 +176,13 @@ parseCLIOpts defaultStreamSize = do
                     Just x -> do
                         s <- evaluate $ getStreamSize x
                         return (s, args)
-                    Nothing -> return (defaultStreamSize, args)
+                    Nothing -> return (defStreamSize, args)
 
     -- Parse gauge options
     let config = defaultConfig
                 { timeLimit = Just 1
                 , minDuration = 0
-                , includeFirstIter = streamSize > defaultStreamSize
+                , includeFirstIter = streamSize > defStreamSize
                 }
     let (cfg, benches) = parseWith config args'
     streamSize `seq` return (streamSize, cfg, benches)
